requirements.txt
python-dotenv
telethon
python-telegram-bot
